+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## [WIP] 0.4.2 / 2018-March-??
 
 - Upgrade Janino version to `3.0.8`
-- [Todo] Add fn to deinit Logback in `util.clj`
+- Add utility fns in `cambium.logback.core.util` namespace
+  - `logger-context-name` to find logger-context name
+  - `stop-logger-context` to stop logger-context
 - Bump cambium.core and cambium.codec-simple dev dependency version to `0.9.2`
   - [Todo] Use the GA version